+++ src/DI/index.ts
@@ -1,0 +1,3 @@
+export { default as Injectable } from './injectable';
+export { default as Injected } from './injected';
+export { injector, factoryCreator } from './factory-creator';
